QueryTest: Update to AssertSql Move AssertBaseline to Relational.Specification.Tests Make AssertSql assert the ordering Add AssertContainsSql which disregard ordering (due to .NET Core) Add output to VS output window again
